refactor(core): consolidate media forwarding on HTMLMediaElementLayer
Move the full Video surface (including presentation APIs) onto
HTMLMediaElementLayer so hosts stay thin typed entry points. Reorder
generics to Target/Events/Next and document why the layer implements Video.
