+++ packages/core/src/dom/media/html-audio-element-host.ts
@@ -1,4 +1,6 @@
 import type { Audio, AudioEvents } from '../../core/media/types';
 import { HTMLMediaElementHost } from './html-media-element-host';
 
-export class HTMLAudioElementHost extends HTMLMediaElementHost<HTMLAudioElement, Audio, AudioEvents> implements Audio {}
+export class HTMLAudioElementHost<Target extends HTMLAudioElement = HTMLAudioElement>
+  extends HTMLMediaElementHost<Target, AudioEvents>
+  implements Audio {}
