+++ packages/core/src/dom/media/html-video-element-host.ts
@@ -1,84 +1,6 @@
-import { isFunction } from '@videojs/utils/predicate';
 import type { Video, VideoEvents } from '../../core/media/types';
-import type { WebKitPresentationMode, WebKitVideoElement } from '../presentation/types';
 import { HTMLMediaElementHost } from './html-media-element-host';
 
-export class HTMLVideoElementHost extends HTMLMediaElementHost<HTMLVideoElement, Video, VideoEvents> implements Video {
-  get poster() {
-    return this.next?.poster ?? '';
-  }
-
-  set poster(value: string) {
-    if (this.next) this.next.poster = value;
-  }
-
-  get playsInline() {
-    return this.next?.playsInline ?? false;
-  }
-
-  set playsInline(value: boolean) {
-    if (this.next) this.next.playsInline = value;
-  }
-
-  get videoWidth() {
-    return this.next?.videoWidth ?? 0;
-  }
-
-  get videoHeight() {
-    return this.next?.videoHeight ?? 0;
-  }
-
-  get disablePictureInPicture() {
-    return this.next?.disablePictureInPicture ?? false;
-  }
-
-  set disablePictureInPicture(value: boolean) {
-    if (this.next) this.next.disablePictureInPicture = value;
-  }
-
-  get webkitPresentationMode() {
-    return (this.next as WebKitVideoElement | null)?.webkitPresentationMode;
-  }
-
-  get webkitSetPresentationMode(): ((mode: WebKitPresentationMode) => void) | undefined {
-    const fn = (this.next as WebKitVideoElement | null)?.webkitSetPresentationMode;
-    return isFunction(fn) ? fn.bind(this.next) : undefined;
-  }
-
-  get isPictureInPicture(): boolean {
-    const { target } = this;
-    return (
-      (!!target && globalThis.document?.pictureInPictureElement === target) ||
-      this.webkitPresentationMode === 'picture-in-picture'
-    );
-  }
-
-  get isFullscreen(): boolean {
-    const { target } = this;
-    if (!target) return false;
-    if (this.webkitPresentationMode === 'fullscreen') return true;
-    const doc = globalThis.document;
-    if (!doc) return false;
-    return (
-      doc.fullscreenElement === target || ('webkitFullscreenElement' in doc && doc.webkitFullscreenElement === target)
-    );
-  }
-
-  async requestPictureInPicture() {
-    if (!this.next) return Promise.reject();
-    return this.next.requestPictureInPicture();
-  }
-
-  async exitPictureInPicture() {
-    return globalThis.document?.exitPictureInPicture();
-  }
-
-  requestFullscreen() {
-    if (!this.next) return Promise.reject();
-    return this.next.requestFullscreen();
-  }
-
-  exitFullscreen() {
-    return globalThis.document?.exitFullscreen();
-  }
-}
+export class HTMLVideoElementHost<Target extends HTMLVideoElement = HTMLVideoElement>
+  extends HTMLMediaElementHost<Target, VideoEvents>
+  implements Video {}
